+++ webapp/src/components/Modals/RentConfirmedModal/CTA/CTA.types.ts
@@ -1,0 +1,5 @@
+export type Props = {
+  to: string
+  name: string
+  isDisabledOnMobile?: boolean
+}
