+++ webapp/src/components/Modals/RentConfirmedModal/RentConfirmedModal.module.css
@@ -1,60 +1,3 @@
-.build-more {
-  background-image: url('../../../images/rentals/build-more.svg');
-}
-
-.get-creative {
-  background-image: url('../../../images/rentals/get-creative.svg');
-}
-
-.manage-land {
-  background-image: url('../../../images/rentals/manage-land.svg');
-}
-
-.icon {
-  width: 52px;
-  height: 48px;
-  background-color: black;
-  background-repeat: no-repeat;
-  background-position: center;
-  border-radius: 5px;
-}
-
-.ctaContainer {
-  display: flex;
-  align-items: center;
-  background-color: rgba(115, 110, 125, 0.24);
-  margin-bottom: 16px;
-  padding: 16px;
-  border-radius: 6px;
-}
-
-.mobileAvailabilityContainer {
-  border: 2px solid #736e7d;
-  border-radius: 6px;
-  margin-bottom: 16px;
-  padding: 13px;
-  display: flex;
-}
-
-.ctaSubtitleMobile {
-  color: #cfcdd4;
-  font-size: 15px;
-  font-weight: 400;
-  opacity: 0.4;
-  margin-top: 2px;
-}
-
-.ctaTextContainer {
-  display: flex;
-  flex-direction: column;
-  margin-left: 16px;
-  color: var(--text);
-}
-
-.ctaSubtitle {
-  color: #cfcdd4;
-}
-
 .textContainer {
   margin-bottom: 32px;
   padding: 0 64px;
@@ -65,17 +8,6 @@
   margin-bottom: 32px;
 }
 
-.infoIcon {
-  width: 16px;
-  height: 16px;
-  background-image: url('../../../images/info.svg');
-  background-repeat: no-repeat;
-  background-position: center;
-  display: inline-block;
-  margin-right: 4px;
-  padding-top: 2px;
-}
-
 @media (max-width: 767px) {
   .textContainer {
     padding: 0;
@@ -83,19 +15,5 @@
 
   :global(.ui.modal > .content) {
     font-size: 15px;
-  }
-
-  .icon {
-    min-width: 48px;
-    min-height: 48px;
-    align-self: flex-start;
-  }
-
-  .ctaTextContainer {
-    font-size: 15px;
-  }
-
-  .ctaContainer {
-    padding: 13px;
   }
 }
